refactor: move parsing to tf pkg
